chore!: rename scores.py -> postprocess.py
